Telecom_AI_Assistant/Telecom_Professional/requirements.txt
streamlit==1.28.1
groq==0.4.2
chromadb==0.4.18
pandas==2.1.3
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
streamlit-lottie==0.0.5
Pillow==10.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openpyxl==3.1.2
python-pptx==0.6.21 